Geographic distribution and access to cancer care in Vietnam
In Vietnam, the National Cancer Control Plan was started in 2008 with aims to provide effective diagnosis and treatment. Even though the Cancer Control Network over 2011-2020 had been approved by the Ministry of Health, the cancer control infrastructure in Vietnam is still inadequate.[1] Given the limited resources, reallocation of available facilities and infrastructure to low-accessibility areas is thus of great importance. 
Therefore, this research was conducted with following objectives:
•	To describe the geographic distribution of cancer care facilities in Vietnam.
•	To estimate the disparity in travel distance to the nearest cancer care facility by provincial level in Vietnam
•	To describe the geographic distribution of radiotherapy facilities in Vietnam and disparity in travel distance to the nearest radiotherapy facility by provincial level
Identifying disparities in geographic access to cancer care will help to locate the shortage areas and inform the cancer control planning and policy to improve the overall accessibility.